Refactor common event aggregation search code into utility class
